requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
opencv-python==4.9.0.80
deepface==0.0.93
numpy==1.26.4
python-multipart==0.0.9
pydantic==2.6.1
requests==2.31.0
pillow==10.2.0